+++ docs/changelog.md
@@ -14,6 +14,9 @@
 - Settings form
 - Markdown support for descripton, note fields
 - game_tags, platform_tags to template_tags
+- Export functions for Lutris, Steam
+- Recently Modified sidebar module
+- scan_games, get_lutris_data, get_steam_data to `extensions.utils`
 
 ### Changed
 
@@ -22,6 +25,7 @@
 - last_played field from datetime to int in `library.models`
 - Added query support for pagination
 - Updated Library UI
+- Reconfigured system directory settings
 
 
 ## [0.0.0] - 2024.0309